+++ packages/core/bootstrap/src/lib/modules/error.ts
@@ -13,6 +13,7 @@ export class AdapterError extends Error {
   errorResponse: unknown
   feedID?: string
   providerStatusCode?: number
+  network?: string
   metricsLabel: HttpRequestType
 
   constructor({
@@ -26,6 +27,7 @@ export class AdapterError extends Error {
     errorResponse,
     feedID,
     providerStatusCode,
+    network,
   }: Partial<AdapterError>) {
     super(message)
 
@@ -43,6 +45,7 @@ export class AdapterError extends Error {
     }
     this.errorResponse = errorResponse
     this.providerStatusCode = providerStatusCode
+    this.network = network
     this.metricsLabel = HttpRequestType.ADAPTER_ERROR
   }
 
@@ -54,6 +57,7 @@ export class AdapterError extends Error {
       url: this.url,
       errorResponse: this.errorResponse,
       feedID: this.feedID,
+      network: this.network,
     }
     const errorFull = { ...errorBasic, stack: this.stack, cause: this.cause }
     return {