+++ packages/composites/dns-record-check/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/dns-record-check-adapter
 
+## 1.3.1
+
+### Patch Changes
+
+- Updated dependencies [6054a7b69]
+- Updated dependencies [816b3d307]
+  - @chainlink/ea-bootstrap@2.12.0
+  - @chainlink/ea-test-helpers@1.4.0
+  - @chainlink/dns-query-adapter@1.5.1
+
 ## 1.3.0
 
 ### Minor Changes