+++ packages/sources/blocksize-capital/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/blocksize-capital-adapter
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [6054a7b69]
+- Updated dependencies [816b3d307]
+  - @chainlink/ea-bootstrap@2.12.0
+  - @chainlink/ea-test-helpers@1.4.0
+
 ## 1.0.11
 
 ### Patch Changes